+++ index.html
@@ -10,6 +10,7 @@
 </head>
 
 <body>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -229,7 +230,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="js\main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
